requirements.txt
db-dtypes
dbt-bigquery
google-cloud-bigquery
pandas
plotly
streamlit
